+++ packages/fast-crud/src/components/crud/fs-row-handle.vue
@@ -25,6 +25,8 @@
                 :is="ui.dropdownItem.name"
                 v-if="item.show !== false && isDropdownBtn(item, index)"
                 :[ui.dropdownItem.command]="item.key"
+                v-bind="item.dropdownItem"
+                :disabled="item.disabled"
               >
                 <div class="fs-row-handle-dropdown-item" v-bind="item">
                   <fs-icon v-if="item.icon" :icon="item.icon" /> {{ item.text || item.title }}